feat: Upgrade dependencies and refine features
This commit updates several dependencies in the `Cargo.toml` file to their latest versions. This
includes `axum`, `tokio`, `clap`, `globmatch`, and others.
Also, it refines the feature set for some dependencies, such as disabling default features for
`serde_json`, `chrono`, `axum`, and `futures`, and enabling specific features like `http1` and
`tokio` for `axum`.
Finally, the `notify` version is updated to 8.0.0.